🚚 api: remove method to a new session repository in src/certification/shared
