+++ api/lib/domain/usecases/get-session-results-by-result-recipient-email.js
@@ -3,10 +3,10 @@ import _ from 'lodash';
 const getSessionResultsByResultRecipientEmail = async function ({
   sessionId,
   resultRecipientEmail,
-  sessionRepository,
+  sharedSessionRepository,
   certificationResultRepository,
 }) {
-  const session = await sessionRepository.getWithCertificationCandidates({ id: sessionId });
+  const session = await sharedSessionRepository.getWithCertificationCandidates({ id: sessionId });
   const certificationCandidateIdsForResultRecipient = _(session.certificationCandidates)
     .filter({ resultRecipientEmail })
     .map('id')
